parinfer-double-quote: invoke parinfer right after insertion
We do not invoke parinfer if point is within a string. This is how it
should be. However, when we're creating a new string, perhaps on a new
line, we want it to be invoked once.

This command is a shortcut for the following action:

1. Insert ";
2. Insert "; (parinfer invoked after self-insert-command outside of a string)
3. Move back one character

(Just ignore the fact that I also made it insert a whitespace in a
convenient place.)

When we use `insert` alone, the parinfer invocation never happens, so
we have to invoke it manually.