Fix tabs breaking indent mode
parinfer.js does this to tabs:

    function onTab(result) {
      if (result.isInCode) {
        result.ch = DOUBLE_SPACE;
      }
    }

which assumes a tab width of 2. parinferlib inherits this:

    (defun parinferlib--on-tab ()
      (when parinferlib--isInCode
        (setq parinferlib--ch parinferlib--DOUBLE_SPACE)))

and this is the reason why previously all tabs are replaced with two
spaces, putting parens in the wrong place if tab width != 2.

I'm not quite sure whether this fix can be upstreamed, so for now I'm
just going to let-bind the "constant" to the correct number of spaces.

---

Alternatively the same thing can be achieved by untabifying the
correct regions at the right times. That seems less elegant than this,
though.